requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pkg-resources==0.0.0
praw==6.4.0
prawcore==1.0.1
python-dotenv==0.10.3
requests==2.22.0
six==1.13.0
update-checker==0.16
urllib3==1.25.7
websocket-client==0.57.0
